+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 ### Added
-- Added the `APIKey` resource and HTTP API support for POST, GET, PATCH and DELETE.
+- Added the `APIKey` resource and HTTP API support for POST, GET, and DELETE.
 
 ### Fixed
 - Opening an already open Bolt database should not cause sensu-agent to hang